+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # 📋 CHANGELOG - EA Multi-Paires Scalping Pro
 
+## Version 27.4.2 (2025-11-09)
+
+### 🐛 Correctif
+- Fix: Dashboard positionné à droite du graphique MT5
+
+---
+
+
+## Version 27.4.1 (2025-11-09)
+
+### 📝 Documentation
+- Doc: Ajout système gestion versions automatique
+
+---
+
+
 ## Version 27.4 (08 Nov 2025)
 
 ### 🔴 CORRECTIFS CRITIQUES